requirements.txt
# Elite AI Software Development Team - CrewAI
crewai>=0.80.0
crewai-tools>=0.14.0

# LLM providers (use at least one)
openai>=1.0.0

# Optional: for web research
# duckduckgo-search>=6.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
